♻️ refactor profile components
